Break cyclic dependencies: extract effect-testcontainers-neo4j, rename effect-vitest-testcontainers
* Rename effect-vitest-testcontainers to effect-vitest-neo4j

The lib only contained Neo4j-specific vitest utilities so the generic name
was misleading.

* Extract effect-testcontainers-neo4j from effect-testcontainers

Move Neo4jTestContainer into a new effect-testcontainers-neo4j package
so that effect-testcontainers remains fully generic with no Neo4j knowledge.

* Inline test helpers in effect-neo4j integration tests

Eliminates the cyclic dependency:
effect-neo4j -> effect-vitest-neo4j -> effect-testcontainers -> effect-neo4j
