+++ packages/effect-testcontainers-neo4j/package.json
@@ -1,19 +1,17 @@
 {
-  "name": "@evryg/effect-neo4j-schema",
+  "name": "@evryg/effect-testcontainers-neo4j",
   "version": "0.0.1",
   "private": true,
   "type": "module",
   "exports": { ".": "./src/index.ts" },
   "dependencies": {
-    "@effect/platform": "^0.96.0",
     "@evryg/effect-neo4j": "workspace:*",
+    "@evryg/effect-testcontainers": "workspace:*",
+    "@testcontainers/neo4j": "^11.12.0",
     "effect": "^3.21.0"
   },
   "devDependencies": {
-    "@effect/platform-node": "^0.106.0",
     "@effect/vitest": "^0.29.0",
-    "fast-check": "^3.23.2",
-    "@evryg/effect-vitest-neo4j": "workspace:*",
     "vitest": "^3.2.1"
   }
 }
