+++ packages/effect-testcontainers-neo4j/src/Neo4jTestContainer.node.integration.test.ts
@@ -1,7 +1,14 @@
 import { layer, expect } from "@effect/vitest"
 import { Effect, Layer } from "effect"
-import { Neo4jClient, UnconfiguredNeo4jClient } from "@evryg/effect-neo4j"
-import { Neo4jConfigFromVitest } from "@evryg/effect-vitest-testcontainers"
+import { Neo4jClient, Neo4jConfig, UnconfiguredNeo4jClient } from "@evryg/effect-neo4j"
+import { inject } from "vitest"
+
+const Neo4jConfigFromVitest: Layer.Layer<Neo4jConfig> = Layer.succeed(Neo4jConfig, {
+  uri: inject("neo4j").uri,
+  user: "neo4j",
+  password: inject("neo4j").password,
+  database: "neo4j",
+})
 
 const TestNeo4j = UnconfiguredNeo4jClient.pipe(Layer.provide(Neo4jConfigFromVitest))
