+++ packages/effect-testcontainers-neo4j/tsconfig.src.json
@@ -4,8 +4,6 @@
     "rootDir": "src",
     "outDir": "outputs/tsc/dist",
     "tsBuildInfoFile": "outputs/tsc/src.tsbuildinfo",
-    // testcontainers is CJS with no exports map — NodeNext cannot resolve the
-    // build/container-runtime subpath import. Use bundler instead.
     "module": "preserve",
     "moduleResolution": "bundler"
   },
@@ -16,5 +14,12 @@
     "src/**/*.test.ts",
     "src/**/*.test-d.ts"
   ],
-  "references": []
+  "references": [
+    {
+      "path": "../effect-neo4j/tsconfig.src.json"
+    },
+    {
+      "path": "../effect-testcontainers/tsconfig.src.json"
+    }
+  ]
 }
