+++ packages/effect-testcontainers/src/index.ts
@@ -1,3 +1,2 @@
 export { acquireContainer } from "./TestContainer.js"
 export { ComposeEnvironment, makeComposeContainer, type ComposeOptions } from "./ComposeContainer.js"
-export { makeNeo4jTestContainer, Neo4jTestContainerLive, type Neo4jTestContainerOptions } from "./Neo4jTestContainer.js"
