+++ packages/effect-vitest-neo4j/src/neo4j-global-setup.ts
@@ -1,7 +1,7 @@
 import { Layer, ManagedRuntime } from "effect"
 import { Neo4jConfig, UnconfiguredNeo4jClient } from "@evryg/effect-neo4j"
 import { ensureSchema } from "@evryg/effect-neo4j"
-import { Neo4jTestContainerLive } from "@evryg/effect-testcontainers"
+import { Neo4jTestContainerLive } from "@evryg/effect-testcontainers-neo4j"
 import { globSync, readFileSync } from "node:fs"
 import type { GlobalSetupContext } from "vitest/node"
